requirements.txt
boto3==1.34.24
openai==1.9.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.12.30